+++ .run/LongQueuePerf.run.xml
@@ -1,11 +1,11 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="LongQueuePerf" type="Application" factoryName="Application" nameIsGenerated="true">
-    <option name="MAIN_CLASS_NAME" value="org.tools4j.mmap.longQueue.impl.LongQueuePerf" />
+  <configuration default="false" name="LongQueuePerf" type="Application" factoryName="Application">
+    <option name="MAIN_CLASS_NAME" value="org.tools4j.mmap.queue.perf.LongQueuePerf" />
     <module name="mmap.mmap-queue.test" />
     <option name="VM_PARAMETERS" value="-Djvm.resource.tracing=false -Ddisable.thread.safety=true -Dagrona.disable.bounds.checks=true" />
     <extension name="coverage">
       <pattern>
-        <option name="PATTERN" value="org.tools4j.mmap.chronicle.longQueue.*" />
+        <option name="PATTERN" value="org.tools4j.mmap.chronicle.*" />
         <option name="ENABLED" value="true" />
       </pattern>
     </extension>
