+++ mmap-queue/src/test/java/org/tools4j/mmap/queue/perf/ChronicleLongQueuePerf.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.tools4j.mmap.chronicle.longQueue;
+package org.tools4j.mmap.queue.perf;
 
 import net.openhft.chronicle.queue.ChronicleQueue;
 import net.openhft.chronicle.wire.WireType;
