+++ mmap-queue/src/test/java/org/tools4j/mmap/queue/perf/LongQueuePerf.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.tools4j.mmap.longQueue.impl;
+package org.tools4j.mmap.queue.perf;
 
 import org.HdrHistogram.Histogram;
 import org.agrona.concurrent.BusySpinIdleStrategy;
